Fix case where incomplete data could result in read
Fixes a bug which would allow an incomplete amount of data to be read
from the incoming buffer when the last slice of data sent ended with a
number, and was not a line continuation.

In the case that a slice of data such as 'end of previous log line^M6'
was sent over the wire, we would process the remaining bits of the
previous log line and then attempt to read the size of the next line.
Because 6 was alone in the buffer, we assumed it was the size of the
next payload which would entirely break our buffer and ruin things for
many millenia.