Switch Kafka library and add ticks
This adds some output every so often to show which partitions this
consumer is processing as well as what the message volume is.

This diff also temporarily switches to a forked version of the kafka-go
library to pull in a fix when consumers are too slow and falling off the
front.